+++ example/libclang-example/generated_bindings.dart
@@ -14,12 +14,12 @@ class LibClang {
 
   /// Free the given string set.
   void clang_disposeStringSet(
-    ffi.Pointer<CXStringSet> set,
+    ffi.Pointer<CXStringSet> set_1,
   ) {
     _clang_disposeStringSet ??= _dylib.lookupFunction<_c_clang_disposeStringSet,
         _dart_clang_disposeStringSet>('clang_disposeStringSet');
     return _clang_disposeStringSet(
-      set,
+      set_1,
     );
   }
 
@@ -3302,11 +3302,11 @@ const int CINDEX_VERSION = 59;
 const String CINDEX_VERSION_STRING = '0.59';
 
 typedef _c_clang_disposeStringSet = ffi.Void Function(
-  ffi.Pointer<CXStringSet> set,
+  ffi.Pointer<CXStringSet> set_1,
 );
 
 typedef _dart_clang_disposeStringSet = void Function(
-  ffi.Pointer<CXStringSet> set,
+  ffi.Pointer<CXStringSet> set_1,
 );
 
 typedef _c_clang_createIndex = ffi.Pointer<ffi.Void> Function(