+++ test/native_test/native_test.dart
@@ -7,6 +7,7 @@ import 'dart:io';
 import 'dart:math';
 
 import 'package:test/test.dart';
+
 import '../test_utils.dart';
 import 'native_test_bindings.dart' as bindings;
 
@@ -62,10 +63,10 @@ void main() {
     });
     test('array-workaround: Order of access', () {
       final struct1 = bindings.getStruct1();
-      int expectedValue = 1;
-      for (int i = 0; i < struct1.ref.data.dimensions[0]; i++) {
-        for (int j = 0; j < struct1.ref.data.dimensions[1]; j++) {
-          for (int k = 0; k < struct1.ref.data.dimensions[2]; k++) {
+      var expectedValue = 1;
+      for (var i = 0; i < struct1.ref.data.dimensions[0]; i++) {
+        for (var j = 0; j < struct1.ref.data.dimensions[1]; j++) {
+          for (var k = 0; k < struct1.ref.data.dimensions[2]; k++) {
             expect(struct1.ref.data[i][j][k], expectedValue);
             expectedValue++;
           }